Chat update / uploaded on host
Chat dates are now synchronized. Meaning all the dates are translated to the user's local timezone. So the messages are sorted out good also.
Fixed dashboard error message.
Uploaded backend to the firebase functions.
Uploaded the website on a hosting with a working serverside.
Now in 'my profile' real statistics are shown (games and points)

Website URL: https://remoteracer.fun/